Add test setup and organize tests
