+++ identity_storage/src/storage/tests/mod.rs
@@ -1,4 +1,6 @@
 // Copyright 2020-2023 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 mod api;
+mod credential_jws;
 mod credential_validation;
+mod test_utils;
